requirements.txt
litellm
google-cloud-aiplatform
python-dotenv
chromadb
google-generativeai
python-dotenv
litellm
numpy
gradio
google-adk
